refactoro ut download button and preview hunt json
